requirements.txt
fastapi>=0.104.1
fastapi-mcp>=0.1.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.5.0
google-cloud-discoveryengine>=0.11.0
google-auth>=2.25.0
httpx>=0.25.2